Extract some common views into WaitlistCommonViews
